+++ csvviz/cli.py
@@ -3,34 +3,14 @@ import json
 from pathlib import Path
 import re
 import sys
-from typing import Any as typeAny, Mapping as typeMapping, NoReturn as typeNoReturn
-from typing import (
-    Dict as typeDict,
-    List as typeList,
-    Tuple as typeTuple,
-    Union as typeUnion,
-)
-from typing import IO as typeIO
-
+from typing import NoReturn as typeNoReturn
 import click
 
-# from csvviz import __version__ as csvviz_version
 from csvviz.cli_utils import clout, clerr, print_version
-
 from csvviz.cmds.info import command as infocommand
 
 
 SUBCOMMAND_PATHS = [p for p in Path("csvviz/cmds/").glob("*.py") if p.name != "info.py"]
-
-
-# def _callback_print_version(ctx, param, value) -> typeNoReturn:
-#     """
-#     https://click.palletsprojects.com/en/3.x/options/#callbacks-and-eager-options
-#     """
-#     if not value or ctx.resilient_parsing:
-#         return
-#     clout(csvviz_version)
-#     ctx.exit()
 
 
 @click.group()
@@ -42,7 +22,9 @@ SUBCOMMAND_PATHS = [p for p in Path("csvviz/cmds/").glob("*.py") if p.name != "i
     help="Print the version of csvviz",
 )
 def apex(**kwargs):
-    """Welcome to csvviz, a command-line tool for producing visualizations using the Vega-lite spec"""
+    """
+    Welcome to csvviz (cvz), a command-line tool for producing visualizations using the Vega-lite spec
+    """
 
     pass